access control for all actions, ability to leave tournament

+++ src/app/lib/actions.ts
@@ -11,6 +11,10 @@ const createTournamentSchema = validation.z.object({
     userID: validation.z.coerce.number().positive(),
 })
 export async function createTournament(_: any, formData: FormData) {
+    const session = await getLoggedInUser()
+    if (!session.userInfo) {
+        return { message: "You must be logged in to create a tournament" }
+    }
     const rawFormData = { name: formData.get("tournamentName"), userID: formData.get('userID') }
     try {
         const data = createTournamentSchema.parse(rawFormData)
@@ -31,9 +35,17 @@ const changeUsernameSchema = validation.z.object({
     username: validation.z.string().min(3).max(20),
 })
 export async function changeUsername(_: any, formData: FormData) {
+    const session = await getLoggedInUser()
+    if (!session.userInfo) {
+        return { message: "You must be logged in to change usernames" }
+    }
+
     try {
         const rawFormData = { username: formData.get("username"), id: formData.get('id') }
         const data = changeUsernameSchema.parse(rawFormData)
+        if (data.id !== session.userInfo.id) {
+            return { message: "You can't change someone else's username" }
+        }
         await db.changeUsername(data.id, data.username)
     } catch (error) {
         if (error instanceof validation.z.ZodError) {
@@ -55,10 +67,19 @@ const inviteToTournamentSchema = validation.z.object({
     email: validation.z.string().email(),
 })
 export async function inviteToTournament(_: any, formData: FormData) {
+    const session = await getLoggedInUser()
+    if (!session.userInfo) {
+        return { message: "You must be logged in to invite someone to a tournament" }
+    }
+
     let data
     try {
         const rawFormData = { userID: formData.get('userID'), tournamentID: formData.get('tournamentID'), tournamentName: formData.get('tournamentName'), email: formData.get('email') }
         data = inviteToTournamentSchema.parse(rawFormData)
+        if (data.userID !== session.userInfo.id) {
+            return { message: "You can't claim to be someone else creating a tournament invite" }
+        }
+        // database checks that data.userID is in the tournament
         await db.addTournamentInvite(data.userID, data.tournamentID, data.email)
         // TODO: email the user an invite link
     } catch (error) {
@@ -79,10 +100,26 @@ const removeInviteSchema = validation.z.object({
     inviteID: validation.z.coerce.number().positive(),
 })
 export async function removeInvite(_: any, formData: FormData) {
+    const session = await getLoggedInUser()
+    if (!session.userInfo) {
+        return { message: "You must be logged in to remove an invite" }
+    }
+
     let data
     try {
         const rawFormData = { tournamentName: formData.get('tournamentName'), inviteID: formData.get('inviteID') }
         data = removeInviteSchema.parse(rawFormData)
+
+        const { tournamentName } = data
+
+        const invite = await db.getTournamentInvite(data.inviteID)
+        const tournaments = await db.getUserTournaments(session.userInfo.id)
+        const userIsInTournament = tournaments.some(t => (t.name === tournamentName))
+        const userIsInvitee = (invite.invitee_email === session.userInfo.email)
+        if (!(userIsInTournament || userIsInvitee)) {
+            return { message: "You aren't authorized to make this change" }
+        }
+
         await db.removeTournamentInvite(data.inviteID)
     } catch (error) {
         if (error instanceof validation.z.ZodError) {
@@ -113,7 +150,11 @@ export async function acceptInvite(_: any, formData: FormData) {
         const rawFormData = { inviteID: formData.get('inviteID'), userID: formData.get('userID') }
         const data = acceptInviteSchema.parse(rawFormData)
         if (data.userID !== session.userInfo.id) {
-            return { message: "You must be logged in to accept an invite" }
+            return { message: "This invite is not for you" }
+        }
+        const invite = await db.getTournamentInvite(data.inviteID)
+        if (invite.invitee_email !== session.userInfo.email) {
+            return { message: "This invite is not for you" }
         }
         await db.acceptTournamentInvite(data.inviteID)
     } catch (error) {
@@ -129,4 +170,37 @@ export async function acceptInvite(_: any, formData: FormData) {
     }
 
     revalidatePath(`/user`)
+}
+
+const leaveTournamentSchema = validation.z.object({
+    tournamentID: validation.z.coerce.number().positive(),
+    tournamentName: validation.z.string().min(3).max(32),
+    userID: validation.z.coerce.number().positive(),
+})
+export async function leaveTournament(_: any, formData: FormData) {
+    const session = await getLoggedInUser()
+    if (!session.userInfo) {
+        return { message: "You must be logged in to leave a tournament" }
+    }
+    let data
+    try {
+        const rawFormData = { tournamentID: formData.get('tournamentID'), tournamentName: formData.get('tournamentName'), userID: formData.get('userID') }
+        data = leaveTournamentSchema.parse(rawFormData)
+        if (data.userID !== session.userInfo.id) {
+            return { message: "You can't leave a tournament for someone else" }
+        }
+        await db.leaveTournament(data.tournamentID, data.userID)
+    } catch (error) {
+        if (error instanceof validation.z.ZodError) {
+            log.error('leaveTournament: validation error: ' + error)
+            return { message: "An error occurred" }
+        } else if (error instanceof db.DBError) {
+            return { message: error.message }
+        }
+        log.error('leaveTournament: unknown error: ' + error)
+        return { message: 'An unknown error occurred' }
+    }
+
+    revalidatePath(`/user`)
+    revalidatePath(`/tournaments/${data.tournamentName}`)
 }
